Fixed script for status report and cleaned up
