+++ scripts/check_xml/Old_xsd(deprecated)/ec.xsd
