+++ scripts/check_xml/exe_package/env.xsd
