+++ scripts/check_xml/restrictive(deprecated)/datas.xsd
