+++ scripts/check_xml/restrictive(deprecated)/ds.xsd
