+++ scripts/check_xml/restrictive(deprecated)/ec.xsd
