+++ scripts/check_xml/restrictive(deprecated)/env.xsd
