+++ scripts/check_xml/restrictive(deprecated)/wsse.xsd
