+++ scripts/check_xml/restrictive(deprecated)/wsu.xsd
