+++ scripts/check_xml/unrestrictive/ec.xsd
