+++ scripts/check_xml/unrestrictive/wsse.xsd
