+++ scripts/guarantee/database_connection.py
