+++ scripts/update-token.py
